+++ packages/client/src/devices/index.ts
@@ -1,1 +1,5 @@
 export * from './devices';
+export * from './InputMediaDeviceManager';
+export * from './InputMediaDeviceManagerState';
+export * from './CameraManager';
+export * from './MicrophoneManager';