+++ packages/react-native-sdk/src/components/Call/CallControls/ToggleVideoPublishingButton.tsx
@@ -1,18 +1,14 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import { AxiosError, OwnCapability } from '@stream-io/video-client';
+import React from 'react';
+import { OwnCapability } from '@stream-io/video-client';
 import {
   Restricted,
   useCall,
-  useHasPermissions,
-  useI18n,
+  useCallStateHooks,
 } from '@stream-io/video-react-bindings';
 import { CallControlsButton } from './CallControlsButton';
 import { muteStatusColor } from '../../../utils';
-import { Alert } from 'react-native';
 import { theme } from '../../../theme';
 import { Video, VideoSlash } from '../../../icons';
-import { usePermissionNotification } from '../../../hooks';
-import { useMediaStreamManagement } from '../../../providers';
 
 /**
  * Props for the Toggle Video publishing button
@@ -31,69 +27,25 @@ export type ToggleVideoPublishingButtonProps = {
 export const ToggleVideoPublishingButton = ({
   onPressHandler,
 }: ToggleVideoPublishingButtonProps) => {
-  const [isAwaitingApproval, setIsAwaitingApproval] = useState(false);
-  const { isVideoMuted, toggleVideoMuted } = useMediaStreamManagement();
-  const { t } = useI18n();
   const call = useCall();
-  const userHasSendVideoCapability = useHasPermissions(
-    OwnCapability.SEND_VIDEO,
-  );
+  const { useCameraState } = useCallStateHooks();
+  const { status } = useCameraState();
 
-  useEffect(() => {
-    if (userHasSendVideoCapability) {
-      setIsAwaitingApproval(false);
-    }
-  }, [userHasSendVideoCapability]);
-
-  const handleRequestPermission = useCallback(
-    async (permission: OwnCapability) => {
-      if (!call?.permissionsContext.canRequest(permission)) {
-        return;
-      }
-      setIsAwaitingApproval(true);
-      try {
-        await call.requestPermissions({ permissions: [permission] });
-      } catch (error) {
-        if (error instanceof AxiosError) {
-          console.log(
-            'RequestPermissions failed',
-            error.response?.data.message,
-          );
-        }
-      }
-    },
-    [call],
-  );
-
-  usePermissionNotification({
-    permission: OwnCapability.SEND_VIDEO,
-    messageApproved: t('You can now share your video.'),
-    messageRevoked: t('You can no longer share your video.'),
-  });
-
-  const handleToggleVideoButton = async () => {
+  const onPress = async () => {
     if (onPressHandler) {
       onPressHandler();
       return;
     }
-    if (userHasSendVideoCapability) {
-      toggleVideoMuted();
-      return;
-    }
-    if (!isAwaitingApproval) {
-      await handleRequestPermission(OwnCapability.SEND_VIDEO);
-    } else {
-      Alert.alert('Awaiting for an approval to share your video.');
-    }
+    await call?.camera.toggle();
   };
 
   return (
     <Restricted requiredGrants={[OwnCapability.SEND_VIDEO]}>
       <CallControlsButton
-        onPress={handleToggleVideoButton}
-        color={muteStatusColor(isVideoMuted)}
+        onPress={onPress}
+        color={muteStatusColor(status === 'disabled')}
       >
-        {isVideoMuted ? (
+        {status === 'disabled' ? (
           <VideoSlash color={theme.light.static_white} />
         ) : (
           <Video color={theme.light.static_black} />