+++ sample-apps/react/react-dogfood/components/CallRecordings.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx';
-import { forwardRef, useEffect, useRef, useState } from 'react';
+import { forwardRef, useCallback, useEffect, useState } from 'react';
 import {
   CallRecording,
   CallRecordingList,
@@ -7,68 +7,51 @@ import {
   MenuToggle,
   ToggleMenuButtonProps,
   useCall,
-  useCallStateHooks,
-  useStreamVideoClient,
 } from '@stream-io/video-react-sdk';
 
-const MAX_NUMBER_POLL_REQUESTS = 6;
-const POLL_INTERVAL_MS = 10 * 1000;
-
 export const CallRecordings = () => {
-  const { useCallRecordings } = useCallStateHooks();
-  const callRecordings = useCallRecordings();
-  const client = useStreamVideoClient();
-  const activeCall = useCall();
-
+  const call = useCall();
+  const [callRecordings, setCallRecordings] = useState<CallRecording[]>([]);
   const [loadingCallRecordings, setLoadingCallRecordings] = useState(false);
-  const recordingPollRequestsCount = useRef(0);
-  const recordingPollRequestsInterval = useRef<
-    ReturnType<typeof setInterval> | undefined
-  >();
+
+  const fetchCallRecordings = useCallback(() => {
+    if (!call) return;
+    call.queryRecordings().then(({ recordings }) => {
+      setCallRecordings(recordings);
+      setLoadingCallRecordings(false);
+    });
+  }, [call]);
 
   useEffect(() => {
-    if (!(client && activeCall && callRecordings)) return;
-    const scheduleCallRecordingPolling = () => {
-      setLoadingCallRecordings(true);
+    fetchCallRecordings();
+  }, [fetchCallRecordings]);
 
-      if (recordingPollRequestsInterval.current) {
-        clearInterval(recordingPollRequestsInterval.current);
-      }
+  useEffect(() => {
+    if (!call) return;
+    const unsubscribeRecordingStopped = call.on('call.recording_stopped', () =>
+      setLoadingCallRecordings(true),
+    );
 
-      recordingPollRequestsInterval.current = setInterval(async () => {
-        let recordings: CallRecording[] = [];
-        try {
-          const response = await activeCall.queryRecordings();
-          recordings = response.recordings;
-        } catch (e) {
-          console.error('Failed to query recordings', e);
-        } finally {
-          recordingPollRequestsCount.current++;
-          const hasNewRecordings = callRecordings.length < recordings.length;
-
-          if (
-            recordingPollRequestsCount.current === MAX_NUMBER_POLL_REQUESTS ||
-            hasNewRecordings
-          ) {
-            clearInterval(recordingPollRequestsInterval.current);
-            setLoadingCallRecordings(false);
-          }
-        }
-      }, POLL_INTERVAL_MS);
-    };
-
-    client.on('call.recording_stopped', scheduleCallRecordingPolling);
+    // @ts-expect-error
+    const unsubscribeRecordingReady = call.on('call.recording_ready', (e) => {
+      // FIXME OL this event isn't yet available in the OpenAPI schema
+      const { call_recording: recording } = e;
+      setCallRecordings((prev) => [...prev, recording]);
+      setLoadingCallRecordings(false);
+    });
 
     return () => {
-      client.off('call.recording_stopped', scheduleCallRecordingPolling);
+      unsubscribeRecordingReady();
+      unsubscribeRecordingStopped();
     };
-  }, [client, activeCall, callRecordings]);
+  }, [call]);
 
   return (
     <MenuToggle placement="bottom-end" ToggleButton={ToggleMenuButton}>
       <CallRecordingList
         callRecordings={callRecordings}
         loading={loadingCallRecordings}
+        onRefresh={fetchCallRecordings}
       />
     </MenuToggle>
   );